backend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
redis==5.0.1
pydantic==2.5.2
python-multipart==0.0.6
